+++ modules/Sources/Dependencies/SDKSynchronizer/SDKSynchronizerTest.swift
@@ -29,7 +29,8 @@ extension SDKSynchronizerClient: TestDependencyKey {
         getSaplingAddress: XCTUnimplemented("\(Self.self).getSaplingAddress", placeholder: nil),
         sendTransaction: XCTUnimplemented("\(Self.self).sendTransaction", placeholder: .placeholder()),
         shieldFunds: XCTUnimplemented("\(Self.self).shieldFunds", placeholder: .placeholder()),
-        wipe: XCTUnimplemented("\(Self.self).wipe")
+        wipe: XCTUnimplemented("\(Self.self).wipe"),
+        switchToEndpoint: XCTUnimplemented("\(Self.self).switchToEndpoint")
     )
 }
 
@@ -50,7 +51,8 @@ extension SDKSynchronizerClient {
         getSaplingAddress: { _ in return nil },
         sendTransaction: { _, _, _, _ in return .placeholder() },
         shieldFunds: { _, _, _ in return .placeholder() },
-        wipe: { Empty<Void, Error>().eraseToAnyPublisher() }
+        wipe: { Empty<Void, Error>().eraseToAnyPublisher() },
+        switchToEndpoint: { _ in }
     )
 
     public static let mock = Self.mocked()
@@ -169,7 +171,8 @@ extension SDKSynchronizerClient {
                 zecAmount: Zatoshi(10)
             )
         },
-        wipe: @escaping () -> AnyPublisher<Void, Error>? = { Fail(error: "Error").eraseToAnyPublisher() }
+        wipe: @escaping () -> AnyPublisher<Void, Error>? = { Fail(error: "Error").eraseToAnyPublisher() },
+        switchToEndpoint: @escaping (LightWalletEndpoint) async throws -> Void = { _ in }
     ) -> SDKSynchronizerClient {
         SDKSynchronizerClient(
             stateStream: stateStream,
@@ -187,7 +190,8 @@ extension SDKSynchronizerClient {
             getSaplingAddress: getSaplingAddress,
             sendTransaction: sendTransaction,
             shieldFunds: shieldFunds,
-            wipe: wipe
+            wipe: wipe,
+            switchToEndpoint: switchToEndpoint
         )
     }
 }